requirements.txt
Flask
python-dotenv
Flask-WTF
WTForms